[ASAA-175] Create Data Module
- Created Data Module
- Moved and reorganized previously created files and classes from `Domain` to `Data`
