+++ data/src/commonMain/kotlin/com/br/kmmdemo/converters/DailyDtoToDailyConverter.kt
@@ -1,4 +1,4 @@
-package com.br.kmmdemo.models.converters
+package com.br.kmmdemo.converters
 
 import com.br.kmmdemo.models.Daily
 import com.br.kmmdemo.network.dtos.DailyDto
