+++ data/src/commonMain/kotlin/com/br/kmmdemo/converters/DailyVDtoToDailyVConverter.kt
@@ -1,4 +1,4 @@
-package com.br.kmmdemo.models.converters
+package com.br.kmmdemo.converters
 
 import com.br.kmmdemo.models.DailyValues
 import com.br.kmmdemo.network.dtos.DailyDataPointValuesDto
