+++ data/src/commonMain/kotlin/com/br/kmmdemo/converters/LocationDtoToLocationConverter.kt
@@ -1,4 +1,4 @@
-package com.br.kmmdemo.models.converters
+package com.br.kmmdemo.converters
 
 import com.br.kmmdemo.models.Location
 import com.br.kmmdemo.network.dtos.LocationDto
