+++ data/src/commonMain/kotlin/com/br/kmmdemo/converters/MinutelyDtoToMinutelyConverter.kt
@@ -1,4 +1,4 @@
-package com.br.kmmdemo.models.converters
+package com.br.kmmdemo.converters
 
 import com.br.kmmdemo.models.Minutely
 import com.br.kmmdemo.network.dtos.MinutelyDto
