+++ data/src/commonMain/kotlin/com/br/kmmdemo/dtos/DailyDto.kt
