+++ data/src/commonMain/kotlin/com/br/kmmdemo/dtos/MinutelyDataPointValuesDto.kt
