+++ data/src/commonMain/kotlin/com/br/kmmdemo/dtos/MinutelyDto.kt
