+++ data/src/commonMain/kotlin/com/br/kmmdemo/network/NetworkRoutes.kt
