+++ data/src/commonMain/kotlin/com/br/kmmdemo/repository/WeatherRepoImplementation.kt
@@ -1,13 +1,14 @@
-package com.br.kmmdemo.repositories
+package com.br.kmmdemo.repository
 
-import com.br.kmmdemo.network.TomorrowIoService
-import com.br.kmmdemo.network.dtos.ForecastDto
+import com.br.kmmdemo.TomorrowIoService
+import com.br.kmmdemo.models.Forecast
+import com.br.kmmdemo.repositories.WeatherRepository
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
 class WeatherRepoImplementation : WeatherRepository, KoinComponent {
     private val tomorrowIoService: TomorrowIoService by inject()
-    override suspend fun getForecastForCity(city: String): Result<ForecastDto> {
+    override suspend fun getForecastForCity(city: String): Result<Forecast> {
         return tomorrowIoService.getForecastByCity(city)
     }
 }
