+++ data/src/commonMain/kotlin/com/br/kmmdemo/useCases/ForecastForCityInteractor.kt
@@ -1,11 +1,10 @@
-package com.br.kmmdemo.usecases.forecastusecase
 
-import com.br.kmmdemo.models.converters.toForecastModel
 import com.br.kmmdemo.repositories.WeatherRepository
-import org.koin.core.component.inject
+import com.br.kmmdemo.usecases.forecastusecase.ForecastForCityUseCase
 import com.br.kmmdemo.usecases.forecastusecase.ForecastForCityUseCase.InvalidCityName
 import com.br.kmmdemo.usecases.forecastusecase.ForecastForCityUseCase.Request
 import com.br.kmmdemo.usecases.forecastusecase.ForecastForCityUseCase.Response
+import org.koin.core.component.inject
 
 class ForecastForCityInteractor: ForecastForCityUseCase {
     private val weatherRepo: WeatherRepository by inject()
@@ -14,6 +13,6 @@ class ForecastForCityInteractor: ForecastForCityUseCase {
         request assertNotNull { forecastRequest ->
             if (forecastRequest.city.isBlank()) throw InvalidCityName()
             val result = weatherRepo.getForecastForCity(forecastRequest.city)
-            Result.success(Response(result.getOrNull().toForecastModel()))
+            Result.success(Response(result.getOrNull()))
         }
 }
