+++ domain/src/commonMain/kotlin/com/br/kmmdemo/TomorrowIoService.kt
@@ -1,0 +1,7 @@
+package com.br.kmmdemo
+
+import com.br.kmmdemo.models.Forecast
+
+interface TomorrowIoService {
+    suspend fun getForecastByCity(city: String): Result<Forecast>
+}
