--- domain/src/commonMain/kotlin/com/br/kmmdemo/network/TomorrowIoService.kt
@@ -1,7 +1,0 @@
-package com.br.kmmdemo.network
-
-import com.br.kmmdemo.network.dtos.ForecastDto
-
-interface TomorrowIoService {
-    suspend fun getForecastByCity(city: String): Result<ForecastDto>
-}
