+++ domain/src/commonMain/kotlin/com/br/kmmdemo/repositories/WeatherRepository.kt
@@ -1,7 +1,7 @@
 package com.br.kmmdemo.repositories
 
-import com.br.kmmdemo.network.dtos.ForecastDto
+import com.br.kmmdemo.models.Forecast
 
 interface WeatherRepository {
-    suspend fun getForecastForCity(city: String): Result<ForecastDto>
+    suspend fun getForecastForCity(city: String): Result<Forecast>
 }
