+++ settings.gradle.kts
@@ -25,3 +25,4 @@ include(":androidApp")
 include(":shared")
 include(":domain")
 include(":compose")
+include(":data")
